# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FreshMood Restaurant QR Ordering API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready, completed, cancelled]
        - name: tableNumber
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Payments
...
components:
  schemas:
...
    OrderItem:
      type: object
      required: [id, orderId, menuItemId, menuItemName, quantity, unitPrice, subtotal]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        menuItemImageUrl:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
        notes:
          type: ["string", "null"]
    OrderItemInput:
      type: object
      required: [menuItemId, quantity]
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
        notes:
          type: string
    Order:
      type: object
      required: [id, customerName, status, subtotal, tax, serviceFee, total, createdAt]
      properties:
        id:
          type: integer
        tableNumber:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        orderType:
          type: string
          enum: [dine_in, take_away, delivery]
        deliveryAddress:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, completed, cancelled]
        subtotal:
          type: number
        tax:
          type: number
        serviceFee:
          type: number
        total:
          type: number
        cashierName:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [unpaid, paid]
        estimatedMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderInput:
      type: object
      required: [customerName, items]
      properties:
        tableNumber:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        notes:
          type: string
        orderType:
          type: string
          enum: [dine_in, take_away, delivery]
        deliveryAddress:
          type: string
        deliveryFee:
          type: integer
          description: Delivery fee chosen by customer (0 for free, 5000 for paid)
        paymentMethod:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderStatusPatch:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, completed, cancelled]
        estimatedMinutes:
          type: integer
...